Add files via upload
I added the typeform in the form as a "start" button at the end of the page because typeform was not giving me a workable code link. This is the link they gave me: <div data-tf-widget="ljfIxEll" data-tf-opacity="83" style="width:100%;height:500px;"></div><script src="//embed.typeform.com/next/embed.js"></script>
So i re worked it into a start button with what had been taught in the previous lessons.